+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- Send a Discord message when a new version is released.
+
 ### 0.32.0
 
 - Removed missiles from tutorial07.